requirements.txt
argcomplete==3.1.2
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Django==4.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-jazzmin==2.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
drf-spectacular==0.26.2
exceptiongroup==1.1.3
fonttools==4.43.1
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
numpy==1.26.1
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.3
Pillow==8.4.0
pipx==1.2.0
pluggy==1.3.0
PyJWT==2.8.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
PySocks==1.7.1
pytest==7.4.3
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
schedule==1.2.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tomli==2.0.1
traitlets==5.12.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
userpath==1.9.1
watchdog==3.0.0
wsproto==1.2.0
